Refactor session engine to delegate pending approval and escalation logic to dedicated functions
