+++ tests/test_session_engine_core.py
@@ -673,97 +673,6 @@ def test_escalation_callback_dispatches_and_clears_notification(tmp_path: Path):
         asyncio.run(_run())
 
 
-def test_pending_requests_include_durable_tool_approvals(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-
-    sid = engine.session_mgr.create_session().session_id
-    engine.session_mgr.append_events(
-        sid,
-        [
-            {
-                "role": "approval_request",
-                "tool_call_id": "tool-1",
-                "tool": "exec",
-                "args": {"command": "date"},
-                "explanation": "run date",
-                "risk_level": "low",
-            }
-        ],
-    )
-    runtime = engine._session_store.load_runtime(sid).model_copy(
-        update={"phase": "waiting_tool_approval", "pending_approval_ids": ["tool-1"]}
-    )
-    engine.session_mgr.save_session_runtime(runtime)
-
-    assert engine.pending_approval_requests(sid) == [
-        {
-            "tool_call_id": "tool-1",
-            "tool": "exec",
-            "args": {"command": "date"},
-            "explanation": "run date",
-            "risk_level": "low",
-        }
-    ]
-
-
-def test_pending_requests_skip_resolved_durable_tool_approvals(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-
-    sid = engine.session_mgr.create_session().session_id
-    engine.session_mgr.append_events(
-        sid,
-        [
-            {"role": "approval_request", "tool_call_id": "tool-1", "tool": "exec", "args": {}},
-            {"role": "approval_response", "tool_call_id": "tool-1", "decision": "approved"},
-        ],
-    )
-    runtime = engine._session_store.load_runtime(sid).model_copy(
-        update={"phase": "waiting_tool_approval", "pending_approval_ids": ["tool-1"]}
-    )
-    engine.session_mgr.save_session_runtime(runtime)
-
-    assert engine.pending_approval_requests(sid) == []
-
-
-def test_pending_requests_include_durable_escalations(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-
-    sid = engine.session_mgr.create_session().session_id
-    engine.session_mgr.append_events(
-        sid,
-        [
-            {
-                "role": "credential_approval",
-                "request_id": "req-1",
-                "vault_paths": ["dev/api"],
-                "names": ["api"],
-                "descriptions": ["API token"],
-                "explanation": "need token",
-            }
-        ],
-    )
-    runtime = engine._session_store.load_runtime(sid).model_copy(
-        update={"phase": "waiting_escalation", "pending_escalation_ids": ["req-1"]}
-    )
-    engine.session_mgr.save_session_runtime(runtime)
-
-    assert engine.pending_escalations(sid) == [
-        {
-            "request_id": "req-1",
-            "kind": "credential_access",
-            "vault_path": "dev/api",
-            "vault_paths": ["dev/api"],
-            "names": ["api"],
-            "descriptions": ["API token"],
-            "skill_name": None,
-            "explanation": "need token",
-        }
-    ]
-
-
 def test_finalize_successful_turn_dispatches_attended_notification(tmp_path: Path):
     async def _run() -> None:
         engine = _make_engine(tmp_path)
